postgres_service/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.27.1
SQLAlchemy==2.0.19
asyncpg==0.27.0
pydantic[email]==1.10.13 
python_dotenv==1.1.0